Teach http_post/3 how to handle JSON bodies
One often wants to POST a JSON body to a web server.  To facilitate
POSTing bodies of different types, the second argument needs to be
wrapped in a compound term.  If the name is 'json', convert the data
structure from a JSON object into codes that can be POSTed.
Otherwise, pass the compound term to http_post_data/3 unchanged.

This required some small changes to those using http_post/3 already.