Teach main/1 how to create new tasks
Any tasks in the Future list (which act as templates for tasks that
will eventually appear in the Today list), that have a due date of
today are automatically copied into the Today list.  The due date is
removed when the new task is created.  That's because once a task
enters the Today list, it stays there until it's finished.  The due
date is no longer relevant.

Semantically, a due date on a task in Future means, "I don't want to
see this task until this date has arrived"