Notes about implementing task deletion
The next goal is to delete template tasks after they've served their
purpose.  Unfortunately, http_open/3 and friends don't support the
HTTP verb DELETE.  To delete a task, we're supposed to make a request
with that verb.  Adding support for DELETE is a good project for my
first contribution to SWI-Prolog itself.  I've created a patch in
~/src/swipl-devel/packages/http that I think will work.  Figure out
how to rebuild my swipl from that patched source.

An alternative might be to send "X-HTTP-Method-Override: DELETE"
header in a POST request.  Google's documentation says this works with
PATCH verb.  I'm not sure about other verbs.  See "Alternate notation
when PATCH HTTP verb is not supported" in
https://developers.google.com/google-apps/tasks/performance