Delete tasks in Future after creating in Today
When a task needs to be moved from the Future list to the Today list,
we must create a new copy of the task in Today (Google has no API for
moving tasks between lists).  After inserting the new copy, we need to
delete the old copy, or we'll copy it the next time too.