Rename weekday to dow
The old name means "day of the week besides Sunday or Saturday" which
isn't actually what I meant.  I meant "day of the week".
Unfortunately, that is really long so I'm settling on the cryptic
"dow".  I can change this later if it proves foolish.  I could also
add day_of_week as an alias, if that makes better sense.