Factor out "one of these days of the week"
This lets one use dow([saturday,sunday]) to constrain a datetime to
only those dates which fell on a Saturday or Sunday.  I had recreated
this same code in two separate clauses of repetition//1 and I'm sure
I'll use it again, so it seemed best to put it in library(julian) for
all to use.