Write TODO test for faulty order dependence
When combining nth/2 and month/1 rules, the order in which they're
applied affects the results.  This shouldn't happen.  I've made a
failing TODO test to remind me to look at this again.  For now, I've
changed repetition//1 to swap the order so that phrases like "second
Tuesday in July" work as expected.