Support "$form during school year"
There are no tests for this form because the phrase "school year" is
different for every jurisdiction.  I need it for my own personal
scheduling, but it won't be useful to others.

This highlights that an English to julian constraints package should
allow users to extend the grammar with additional rules.  We can all
share the universally useful ones and add our own peculiarities on
top.