Implement datetime math via delta_time/3
This predicate describes the duration that passes between two
datetimes.  I'm not happy with using label/1 to avoid the rounding
ambiguity.  I suspect this will cause me troubles later and need to be
removed.  However, it does address all the use cases I care about at
the moment and the API is correct so I should be able to fix it later
without hassle.